pyproject.toml
[project]
name = "creative-approval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "numpy>=2.3.2",
    "openai>=1.106.1",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
]
